Add a CNI focussed variants of kubelet-density (#152)
This commit adds two new kinds of kubelet-density like
workloads that are focussed on CNI performance.

1. kubelet-density-cni: This creates deployments of node.js
webservers fronted by services and another deployment of pods that
run curl against the webserver service for readiness probe. This ensures
that the curl pods only go into ready when they are able to reach the
webserver srvice. kube-burner accordingly accounts for this delay in the
podready metrics.

2. kubelet-density-cni-networkpolicy: This is similar to above workload
except  a deny-all policy is created in the namespace to begin with and
later a newotk policy to allow http on 8080 is introduced when the
pods/services are created. This also ensures that the curl pods'
readiness probe only succeeds when the flows dictated by the
network policy are applied. kube-burner podready metrics now account for
propagation of this network policy.

Closes: #151
Signed-off-by: Sai Sindhur Malleni <smalleni@redhat.com>