Use LastTransitionTime field in pod latency measurement
Some issues have recently popped up with respect to the way kube-burner
calculates latencies of the pod phases.

The original implementation used the current timestamp when the pod was updated
and detected by the Watcher, it turns out that this watcher is not 100% precise and
can lead into delays when the load is high.

Using the LastTransitionTime field of the pod condition fixes the issue,
however it comes with several trade-offs, most remarkable one is that:

- This field is provided in seconds, hence we'll loose the microsecond
  granularity when calculating the different latencies.

Signed-off-by: Raul Sevilla <rsevilla@redhat.com>