Add network-policy workload (#502)
kube-burned modification: add BeforeCleanup param to jobConfig. This
allow to set up a command that should be executed before the workload
is deleted. It is used by network policy workload, since network
policy doesn't have a status, and extra means to track when the config
is applied are needed.
Add Binomial, IndexToCombination and GetSubnet24 functions for
templating, since they are needed to generate complex network policy
configs.

Signed-off-by: Nadia Pinaeva <npinaeva@redhat.com>