Copy the labels to the underlying template even if it does not have any (#737)
## Type of change

- [x] Bug fix

## Description

kube-burner copies the labels of its created objects into the template
of the underlying object (e.g. pod template of a deployment). These
labels are then used in other places in the code. For example, metrics
measurements rely on the existence of the `kube-burner-runid` label on
the pod (or VMI in case of VMs).

In the original code, the labels were copied only if the labels field
already existed.

With these fix, for specific kinds, the labels are copied even if none
were added.

## Related Tickets & Documents

- Closes #736

Signed-off-by: Ygal Blum <ygal.blum@gmail.com>