Additional vars have prececence over env vars (#949)
## Type of change

- Bug fix
- Optimization

## Description

Environment variables now get overwritten by additionalVars, (which is
more logical), that are can be passed from flags in wrappers and
arbitrary YAML files.

This is required to avoid issues with variables using data types
different than "string", which can be overridden by environment
variables

Also removing the variable inputData, which is useless.

## Related Tickets & Documents

- Related Issue #
- Closes #

---------

Signed-off-by: Raul Sevilla <rsevilla@redhat.com>