Stop all watchers and wait for cache sync before collecting the measurements (#1025)
## Type of change

- Bug fix

## Description

Thanks to a flacky test that can be observed at
https://github.com/kube-burner/kube-burner/actions/runs/19163907877/job/54780056514?pr=1024#step:5:137,
its been detected that there're situations where no documents are
indexed by the podLatency measurement.

When stopping the pod latency measurement it's possible that the pod
watcher hasn't received the pod yet, for that reason we should normalize
the metrics after stopping the watcher.


## Related Tickets & Documents

- Related Issue #
- Closes #

---------

Signed-off-by: Raul Sevilla <rsevilla@redhat.com>