Add space for service latency log to play nicer with double-click (#1050)
## Type of change

<!-- Choose a type of change -->

- Refactor

## Description

Service latencies are now printed with no space between the measurement
and the unit ('ms'). When the string is double-clicked, say to copy in
order to paste somewhere else, it includes the 'ms'.

To compare with our other measurements, pod latency units are not
printed. This change makes the behavior slightly more consistent.

Signed-off-by: Andrew Collins <ancollin@redhat.com>