Fix namespaces churning logic (#1061)
## Type of change

- Bug fix

## Description

The logic of namespaces churning is broken. The order of namespaces
returned from List() is not guaranteed, so it's possible for kube-burner
to mistakenly try to re-create a namespace that has not been deleted.

Rolling back to the previous logic that dynamically generates the
namespace name to churn.

It's possible that the issue wasn't detected by CI b/c the number
namespaces to churn was very low (20% of 5 iterations)

## Related Tickets & Documents

- Related Issue #
- Closes #

---------

Signed-off-by: Raul Sevilla <rsevilla@redhat.com>